requirements.txt
Django==1.9.1
gunicorn==19.4.5
psycopg2==2.6.1
djangorestframework==3.3.1
